Re #21: Contrafact discovery is now available.
Contrafact discovery itself is implemented by the ContrafactService, in the discover() method. For now, it returns only a single combined ContrafactSet that contains an Alignment of all the chants that are part of at least one contrafact candidate pair. (There should later be more structure to this.)

The UI for contrafact discovery is to reduce the alignment on screen to just the chants that are in contrafact pairs. Implemented in restrictToContrafacts().

In relation to implementing contrafact discovery, the following things were also done:

- Refactored how distance matrix computation settings are applied into the DistanceService.
- The Alignment model has a new method that creates a subset alignment based on a list of URLs.
- The Alignment model is retrieved differently: it is re-constructed from the data, rather than directly loading the object as JSON. This was done in order to keep older saved alignments working after adding methods to the Alignment class.